Phase 9 turn 4 smoke fix #2: TripPayment.amount is integer cents, not dollars
Manual smoke surfaced a 100x display bug after the receipt-schema
fix landed and the tipDriver Cloud Function patch deployed. The
receipt rendered, the tip flow worked, but a $5 fare and $5 tip
displayed as $500.00 each ($1000.00 total).

Root cause: src/data/mappers/tripPaymentMapper.ts:43 read the wire
`doc.amount` field via Money.fromMajor(doc.amount, 'USD') —
treating it as dollars. But the Stripe webhook server
(yeride-stripe-server/stripe/routes.js:132) writes the raw
`pi.amount` from the Stripe Charge/PaymentIntent, which is ALWAYS
integer cents (Stripe API contract — amounts in the smallest
currency unit). The webhook also writes a sibling `amountInDollars`
field for clients that prefer dollars.

The exact 100x: amount=500 (cents = $5.00) → Money.fromMajor(500,
'USD') → Math.round(500 * 100) = 50000 minor units → format() =
'$500.00'.

The DTO comment claiming "PLAIN NUMBER IN DOLLARS" was a
documentation lie — written under the assumption that the legacy
schema put dollars on disk, but the Stripe webhook (the only writer
to this collection) actually writes cents.

Fix: single-shape extension at the DTO + mapper layer.

  - src/data/dto/TripPaymentDoc.ts: amount field comment rewritten
    to say "INTEGER CENTS — Stripe-native unit." Schema gains
    .int() to enforce the contract — a non-integer here would be a
    wire-format break worth surfacing as a parse failure rather
    than silently truncating. The amountInDollars sibling on the
    wire stays unmapped (cents-as-integer is the source of truth).

  - src/data/mappers/tripPaymentMapper.ts: Money.fromMajor →
    Money.create. Money.create is the integer-minor-units
    constructor; passing cents directly produces the right Money
    value. JSDoc rewritten to document the Stripe-native cents
    contract + warn against the reverted fromMajor path.

Tests: +3 in tripPaymentMapper.test.ts.
  - rejects a non-integer amount (Stripe always writes integer cents)
  - regression: amount=500 (cents) renders as '$5.00' (cents-not-
    dollars contract) — the headline regression
  - regression: tip amount=500 (cents) renders as '$5.00'
Existing fixture amounts (e.g. 12.34) updated to integer cents
(e.g. 1234) so they conform to the new schema.

Why this is safe:
  - All other consumers of TripPayment.amount work at the domain
    layer; Money arithmetic operates on minor units internally.
  - useRideReceiptViewModel's fareTotal math is correct regardless
    of how the wire was interpreted (the bug was contained at the
    parse boundary).
  - seedPayments on InMemoryRideRepository takes domain TripPayment
    values directly — no wire-format assumption.
  - The tip-flow submit path (useTipFlowViewModel → ProcessTip →
    tipDriver Cloud Function) sends a plain dollar amount up the
    wire; the deployed callable expects dollars and multiplies by
    100 internally. Only the read-back path was broken.

Acceptance: 180 suites / 1532 tests passing (+0 suites / +3 tests
over smoke fix #1's 180/1529). Typecheck + lint + format + test
all green.

No native rebuild required. No new dependencies. No plugin patches.

After this lands, the receipt renders fare and tip amounts
correctly: $5 fare → $5.00, $5 tip → $5.00, total $10.00.

+++ src/data/dto/TripPaymentDoc.ts
@@ -8,7 +8,14 @@ import { z } from 'zod';
  * Field semantics:
  *   - `type`: 'fare' (the ride fare), 'tip' (rider's optional gratuity),
  *     'refund' (cancellation refund or post-trip dispute resolution).
- *   - `amount`: PLAIN NUMBER IN DOLLARS. Mapped to `Money` USD minor units.
+ *   - `amount`: INTEGER CENTS — Stripe-native unit. The webhook server
+ *     (yeride-stripe-server/stripe/routes.js:132) writes the raw
+ *     `pi.amount` from the Stripe Charge/PaymentIntent, which is always
+ *     an integer in the smallest currency unit. Maps to `Money` via
+ *     `Money.create(amount, 'USD')` (NOT `fromMajor` — that's a 100x
+ *     bug we hit during the Phase 9 turn 4 smoke). The webhook also
+ *     writes a sibling `amountInDollars` field which the rewrite
+ *     ignores (cents-as-integer is the source of truth).
  *   - `status`: Stripe-derived terminal state for the payment row —
  *     'succeeded', 'failed', or 'refunded'.
  *   - `createdAt`: ISO string. Used for ordering on the receipt screen
@@ -17,7 +24,11 @@ import { z } from 'zod';
 
 export const TripPaymentDocSchema = z.object({
   type: z.enum(['fare', 'tip', 'refund']),
-  amount: z.number().finite().gte(0),
+  // Integer cents (Stripe-native). `.int()` enforces the contract — Stripe
+  // amounts are ALWAYS integers in the smallest currency unit; a non-
+  // integer here would be a wire-format break worth surfacing as a parse
+  // failure rather than silently truncating.
+  amount: z.number().int().finite().gte(0),
   status: z.enum(['succeeded', 'failed', 'refunded']),
   createdAt: z.string().min(1),
 });
